Minor format specifier cross-platform fixes

+++ Common/Include/minibatchsourcehelpers.h
@@ -89,7 +89,7 @@ public:
                     retries++;
                 }
             }
-            fprintf (stderr, "randomordering: %zu retries for %zu elements (%.1f%%) to ensure window condition\n", retries, map.size(), 100.0 * retries / map.size());
+            fprintf(stderr, "randomordering: %lu retries for %lu elements (%.1f%%) to ensure window condition\n", (unsigned long)retries, (unsigned long)map.size(), 100.0 * retries / map.size());
             // ensure the window condition
             foreach_index (t, map) assert ((size_t) t <= map[t] + randomizationrange/2 && map[t] < (size_t) t + randomizationrange/2);
 #if 0       // and a live check since I don't trust myself here yet
